backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic[email]==2.6.0
python-dotenv==1.0.0